requirements.txt
customtkinter
requests
pillow
google-auth
google-api-python-client
google-auth-oauthlib
selenium
spotipy
playsound
azure-cognitiveservices-speech
datasets
transformers[torch]
numpy
torch
pyaudio
pydub
comtypes
pycaw
torchaudio
soundfile